requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.8
requests==2.32.3
urllib3==2.0.7
websockets==12.0

firebase-admin==6.4.0
pycryptodome==3.19.0
openai==1.3.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
pytz==2024.1
# Email service dependencies
python-dotenv==1.0.0
# Email automation dependencies
guardrails-ai==0.4.0
pydantic==2.4.2
aiohttp==3.9.1
